+++ src/booking/domain/model/__shared/aggregate-root.ts
