+++ src/booking/domain/model/__shared/entity-like.ts
