+++ src/booking/domain/model/__shared/entity.ts
