+++ src/booking/domain/model/__shared/event/idomain_event.ts
