+++ src/booking/domain/model/__shared/index.ts
