+++ src/booking/domain/model/__shared/value-object.ts
