+++ src/booking/domain/model/room/room.ts
@@ -1,4 +1,4 @@
-import { Entity, Identifier } from '../../shared';
+import { Entity, Identifier } from '../__shared';
 
 export class RoomId extends Identifier<string> {}
